+++ src/apps/toolbar/modules/main/ContextMenu.tsx
@@ -1,11 +1,13 @@
-import { Menu, Popover } from 'antd';
+import { Flex, Menu, Popover } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
+import { invoke, SeelenCommand } from 'seelen-core';
 
 import { BackgroundByLayersV2 } from 'src/apps/seelenweg/components/BackgroundByLayers/infra';
 
 import { RootActions, Selectors } from '../shared/store/app';
 import { SavePlaceholderAsCustom } from './application';
+import { Icon } from 'src/apps/shared/components/Icon';
 
 export function MainContextMenu() {
   const plugins = useSelector(Selectors.plugins);
@@ -41,9 +43,29 @@ export function MainContextMenu() {
                   </BackgroundByLayersV2>
                 }
               >
-                <div>{t('context_menu.add_module')}</div>
+                <Flex justify="space-between" align="center">
+                  {t('context_menu.add_module')}
+                  <Icon iconName="FaChevronRight" size={12} />
+                </Flex>
               </Popover>
             ),
+          },
+          {
+            type: 'divider',
+          },
+          {
+            key: 'task_manager',
+            label: t('context_menu.task_manager'),
+            onClick() {
+              invoke(SeelenCommand.OpenFile, { path: 'C:\\Windows\\System32\\Taskmgr.exe' });
+            },
+          },
+          {
+            key: 'settings',
+            label: t('context_menu.settings'),
+            onClick() {
+              invoke(SeelenCommand.ShowAppSettings);
+            },
           },
         ]}
       />